feat: get cards and profile info

+++ src/scripts/components/card.js
@@ -1,17 +1,20 @@
 const cardTemplate = document.querySelector("#card-template").content.querySelector(".places__item");
 
-export function createCardElement(data, onDelete, onLike, onImageClick) {
+export function createCardElement(data, onDelete, onLike, onImageClick, userId, cardUserId) {
   const cardElement = cardTemplate.cloneNode(true);
-  const deleteButton = cardElement.querySelector(".card__delete-button");
   const likeButton = cardElement.querySelector(".card__like-button");
   const cardTitle = cardElement.querySelector(".card__title");
   const cardImage = cardElement.querySelector(".card__image");
+
+  if (userId === cardUserId) {
+    const deleteButton = cardElement.querySelector(".card__delete-button");
+    deleteButton.addEventListener("click", () => onDelete(cardElement) );
+  }
 
   cardImage.src = data.link;
   cardImage.alt = data.name;
   cardTitle.textContent = data.name;
 
-  deleteButton.addEventListener("click", () => onDelete(cardElement) );
   likeButton.addEventListener('click', () => onLike(likeButton));
   cardImage.addEventListener('click', () => onImageClick(data.name, data.link))
